# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JISHLink Workforce & Manpower Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily attendance report data
      security:
        - AdminBearerAuth: []
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - name: recruiterId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Daily report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReportResponse"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly attendance report data
      security:
        - AdminBearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - name: recruiterId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Monthly report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReportResponse"
  /reports/export/excel:
    get:
      operationId: exportExcelReport
      tags: [reports]
      summary: Export attendance report as Excel file
      security:
        - AdminBearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [daily, monthly]
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - name: recruiterId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /reports/export/pdf:
    get:
      operationId: exportPdfReport
      tags: [reports]
      summary: Export attendance report as PDF file
      security:
        - AdminBearerAuth: []
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [daily, monthly]
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: query
          schema:
            type: string
            format: uuid
        - name: recruiterId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
    EmployeeBearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DailyReportResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        total:
          type: integer
        present:
          type: integer
        absent:
          type: integer
        late:
          type: integer
        early_out:
          type: integer
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
      required:
        - date
        - total
        - present
        - absent
        - late
        - early_out
        - records
    MonthlyReportResponse:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        working_days:
          type: integer
        records:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyEmployeeRecord"
      required:
        - year
        - month
        - working_days
        - records
    AttendanceRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        employee_code:
          type: string
        employee_name:
          type: string
        site_id:
          type: string
          format: uuid
        site_name:
          type: string
        date:
          type: string
          format: date
        check_in_time:
          type: string
          format: date-time
          nullable: true
        check_out_time:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [present, absent, late, early_out]
        check_in_lat:
          type: number
          nullable: true
        check_in_lng:
          type: number
          nullable: true
      required:
        - id
        - employee_id
        - employee_code
        - employee_name
        - site_id
        - site_name
        - date
        - status
    MonthlyEmployeeRecord:
      type: object
      properties:
        employee_id:
          type: string
          format: uuid
        employee_code:
          type: string
        employee_name:
          type: string
        present_days:
          type: integer
        absent_days:
          type: integer
        late_days:
          type: integer
        attendance_percentage:
          type: number
          format: double
      required:
        - employee_id
        - employee_code
        - employee_name
        - present_days
        - absent_days
        - late_days
        - attendance_percentage
...